requirements.txt
anyio==3.7.0
APScheduler==3.6.3
asgiref==3.7.2
certifi==2023.5.7
crispy-bootstrap5
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.1
django-admin-tools==0.9.2
django-crispy-forms
django-import-export
django-js-asset
django-mptt==0.14.0
django-mptt-admin==2.5.1
et-xmlfile==1.1.0
future==0.18.3
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
lxml==4.9.2
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
Pillow==9.5.0
playsound==1.3.0
pypng==0.20220715.0
python-docx==0.8.11
pytz==2023.3
PyYAML==6.0
qrcode==7.4.2
reportlab==4.0.4
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
tablib==3.5.0
tornado
tqdm==4.65.0
typing_extensions==4.6.3
tzlocal==5.0.1
urllib3==2.0.3
xlrd==2.0.1
xlwt==1.3.0
APScheduler==3.6.3
billiard==3.6.4.0
celery==5.1.2
django-admin-list-filter-dropdown==1.0.2
django-celery-results==2.2.0
django-celerybeat-status==0.0.10
django-celery-beat
django-environ==0.4.5
django-json-widget==1.1.1
ipython==8.8.0
kombu==5.1.0
psycopg2-binary==2.9.3
redis==3.3.11
uwsgi==2.0.20
python-telegram-bot==13.15
telegram_django_bot==1.1.0